% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_create_tags}
\alias{workspaces_create_tags}
\title{Creates the specified tags for the specified WorkSpace}
\usage{
workspaces_create_tags(ResourceId, Tags)
}
\arguments{
\item{ResourceId}{[required] The identifier of the WorkSpace. To find this ID, use
DescribeWorkspaces.}

\item{Tags}{[required] The tags. Each WorkSpace can have a maximum of 50 tags.}
}
\description{
Creates the specified tags for the specified WorkSpace.
}
\section{Request syntax}{
\preformatted{workspaces$create_tags(
  ResourceId = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
