% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_create_workspaces}
\alias{workspaces_create_workspaces}
\title{Creates one or more WorkSpaces}
\usage{
workspaces_create_workspaces(Workspaces)
}
\arguments{
\item{Workspaces}{[required] The WorkSpaces to create. You can specify up to 25 WorkSpaces.}
}
\description{
Creates one or more WorkSpaces.
}
\details{
This operation is asynchronous and returns before the WorkSpaces are
created.
}
\section{Request syntax}{
\preformatted{workspaces$create_workspaces(
  Workspaces = list(
    list(
      DirectoryId = "string",
      UserName = "string",
      BundleId = "string",
      VolumeEncryptionKey = "string",
      UserVolumeEncryptionEnabled = TRUE|FALSE,
      RootVolumeEncryptionEnabled = TRUE|FALSE,
      WorkspaceProperties = list(
        RunningMode = "AUTO_STOP"|"ALWAYS_ON",
        RunningModeAutoStopTimeoutInMinutes = 123,
        RootVolumeSizeGib = 123,
        UserVolumeSizeGib = 123,
        ComputeTypeName = "VALUE"|"STANDARD"|"PERFORMANCE"|"POWER"|"GRAPHICS"|"POWERPRO"|"GRAPHICSPRO"
      ),
      Tags = list(
        list(
          Key = "string",
          Value = "string"
        )
      )
    )
  )
)
}
}

\keyword{internal}
